requirements.txt
# Dev environment
pip
autopep8

# App
flask==2.3.3
google-generativeai
python-dotenv